added jwt

+++ project/app/user/views.py
@@ -3,7 +3,10 @@ from extensions import app, Session
 from flask import request, jsonify
 from .models import User
 from .schema import UserSchema, UserSchemaUpdate
+from .schema import UserLoginSchema
 from imports import *
+from admin_dec import *
+
 
 @app.route('/user', methods=['POST', 'GET'])
 def create_user():
@@ -40,36 +43,77 @@ def create_user():
             return UserSchema(many=True).jsonify(users)
 
 
-
-@app.route('/user/<id>', methods=['GET', 'PUT', 'DELETE'])
-def user(id):
+@app.route('/user/<id>', methods=['GET'])
+@jwt_required()
+@admin_required
+def user_get(id):
     session = Session()
     user_data = session.query(User).get(id)
 
     if user_data:
         if request.method == 'GET':
             return UserSchema().jsonify(user_data)
-
-        elif request.method == 'PUT':
-            user_request = UserSchemaUpdate().load(request.json)
-
-            user_data.first_name = user_request['first_name']
-            user_data.last_name = user_request['last_name']
-
-            if user_request['first_name'] != "" and \
-                    user_request['last_name'] != "":
-                session.commit()
-                return UserSchemaUpdate().jsonify(user_data)
-            else:
-                content = jsonify({"Error": "Fields can`t be empty"})
-                return content, 400, {'content-type': 'application/json'}
-
-        elif request.method == 'DELETE':
-            session.delete(user_data)
-            session.commit()
-            content = jsonify({"Success": "User was deleted"})
-            return content, 200, {'content-type': 'application/json'}
-
     else:
         content = jsonify({"Error": "User doesn't exists"})
         return content, 404, {'content-': 'application/json'}
+
+
+@app.route('/user/<id>', methods=['PUT', 'DELETE'])
+@jwt_required()
+# @admin_required
+def user(id):
+    session = Session()
+    current_identity_email = get_jwt_identity()
+    user_data = session.query(User).get(id)
+
+
+    if user_data:
+        if current_identity_email == user_data.email:
+            if request.method == 'PUT':
+                user_request = UserSchemaUpdate().load(request.json)
+
+                user_data.first_name = user_request['first_name']
+                user_data.last_name = user_request['last_name']
+
+                if user_request['first_name'] != "" and \
+                        user_request['last_name'] != "":
+                    session.commit()
+                    return UserSchemaUpdate().jsonify(user_data)
+                else:
+                    content = jsonify({"Error": "Fields can`t be empty"})
+                    return content, 400, {'content-type': 'application/json'}
+
+            elif request.method == 'DELETE':
+                session.delete(user_data)
+                session.commit()
+                content = jsonify({"Success": "User was deleted"})
+                return content, 200, {'content-type': 'application/json'}
+        else:
+            content = jsonify({"Error": "Access is denied"})
+            return content, 404, {'content-': 'application/json'}
+    else:
+        content = jsonify({"Error": "User doesn't exists"})
+        return content, 404, {'content-': 'application/json'}
+
+
+@app.route('/user/login', methods=['GET'])
+def login_user():
+    session = Session()
+    try:
+        user_request = UserLoginSchema().load(request.json)
+    except ValidationError as e:
+        rv = dict({'message': e.normalized_messages()})
+        return rv, 400, {'content-type': 'application/json'}
+
+    if not user_request:
+        content = jsonify({"Error": "could not verify user_request"})
+        return content, 401, {'content-': 'application/json'}
+
+    user = session.query(User).filter_by(email=user_request['email']).first()
+
+    if user and check_password_hash(user.password, user_request['password']):
+        access_token = create_access_token(identity=user.email)
+        return jsonify({'token': access_token})
+
+    content = jsonify({"Error": "could not verify"})
+    return content, 401, {'content-': 'application/json'}
